requirements-test.txt
coverage
mock>=1.0.1
flake8>=2.1.0
factory_boy==2.5.2
pytest==2.7.3
pytest-cov==2.2.0
pytest-django==2.9.1
pytest-capturelog==0.7
fake-factory==0.5.3
